disable animation on switch, between timeline and barchart, it creates flickering
