+++ src/timeline.ts
@@ -30,9 +30,6 @@ export const defaultTimelineOptions: TimelineOptions = {
 
 export class Timeline extends Chart {
   protected options: TimelineOptions;
-
-  // private ogma: Ogma<any, any>;
-  private isChangingRange: boolean;
   private itemToNodes: Lookup<Id[]>;
 
   /**
@@ -102,31 +99,11 @@ export class Timeline extends Chart {
   }
 
   protected onRangeChange() {
-    // prevent from infinite loop: setdata and window trigger this event
-    if (this.isChangingRange) return;
-
-    // this.isChangingRange = true;
-    const { start, end } = this.chart.getWindow();
-    const length = +end - +start;
-    // choose a scale adapted to zoom
-    const { scale } = scales.reduce(
-      (scale, candidate) => {
-        const bars = Math.round(length / candidate);
-        if (bars < scale.bars && bars > 10) {
-          return {
-            bars,
-            scale: candidate,
-          };
-        }
-        return scale;
-      },
-      { bars: Infinity, scale: Infinity }
-    );
+    const scale = this.getScale();
     if (scale === this.currentScale) {
-      this.isChangingRange = false;
+      return;
     }
     this.emit(scaleChange, { scale, tooZoomed: false });
-    this.isChangingRange = false;
   }
 
   highlightNodes(nodes: NodeList | Id[]) {
